EDAC/igen6: Add registration APIs for In-Band ECC error notification
The igen6_edac driver is the root to capture the In-Band ECC error
event. There are some external modules which want to be notified about
the In-Band ECC errors for specific error handling. So add the
registration APIs for those external modules for the In-Band ECC errors.

Intel-SIG: EDAC/igen6: Add registration APIs for In-Band ECC error notification.
https://github.com/intel/linux-intel-lts/commit/da1536bdff36

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>
[ Qingdong Shi: amend commit log ]
Signed-off-by: Qingdong Shi <qingdong.shi@intel.com>