x86/cpu/hygon: Clear SME feature flag when not in use
hygon inclusion
category: feature
CVE: NA

---------------------------

The commit 08f253ec3767 ("x86/cpu: Clear SME feature flag when not in
use") will clear SME feature flag if the kernel is not using it on AMD
CPUs, this will help userspace to determine if SME is available and in
use from /proc/cpuinfo.

Apply this change to Hygon CPUs as well.

Signed-off-by: hanliyang <hanliyang@hygon.cn>