misc: ntsync: mark driver as "broken" to prevent from building
[ Upstream commit f5b335dc025cfee90957efa90dc72fada0d5abb4 ]

The ntsync code is only partially enabled in the kernel at this point in
time, creating the device node and that's about it.  Don't confuse
systems that expect to see a working ntsync interface by teasing it with
this basic structure at this point in time, so mark the code as "broken"
so that it is not built and enabled just yet.

Once the rest of the code is accepted, this will be reverted so that the
driver can be correctly built and used, but for now, this is the safest
way forward.

Reviewed-by: Elizabeth Figura <zfigura@codeweavers.com>
Link: https://lore.kernel.org/r/2024051450-abrasion-swizzle-550b@gregkh
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[ Backport from v6.10 ]
Signed-off-by: WangYuli <wangyuli@uniontech.com>
(cherry picked from commit 95748f31330941127cec17986428ba7a5eb348b8)
(cherry picked from commit f3ce22c72e9a2c8a64f815488eba90e3165337f8)