ntsync: Introduce NTSYNC_IOC_CREATE_MUTEX.
[ Upstream commit 5bc2479a3585be35ea32e431bc82282bbf5a5d5b ]

This corresponds to the NT syscall NtCreateMutant().

An NT mutex is recursive, with a 32-bit recursion counter. When acquired via
NtWaitForMultipleObjects(), the recursion counter is incremented by one. The OS
records the thread which acquired it.

The OS records the thread which acquired it. However, in order to keep this
driver self-contained, the owning thread ID is managed by user-space, and passed
as a parameter to all relevant ioctls.

The initial owner and recursion count, if any, are specified when the mutex is
created.

Signed-off-by: Elizabeth Figura <zfigura@codeweavers.com>
Link: https://lore.kernel.org/r/20241213193511.457338-6-zfigura@codeweavers.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[ Backport from v6.14 ]
Signed-off-by: WangYuli <wangyuli@uniontech.com>
(cherry picked from commit 200fa7574efe82081bb97ebb0df705a2aa2a7e89)
(cherry picked from commit d51defacff3d8cfebe882f92cb9d11954d02fffa)