sw64: fix for executing syscall(-1)
If a task executes syscall(-1), we intercept this early and force r0 to
be -ENOSYS so that we don't need to distinguish this scenario from one
where the nr is -1 because a tracer wants to skip the system call using
ptrace. With the return value set, the return path is the same as the
skip case.

Signed-off-by: Gu Zitao <guzitao@wxiat.com>
Reviewed-by: He Sheng <hesheng@wxiat.com>
(cherry picked from commit 4371acdf52212bcf9b0d786d45a9496ad375c549)
(cherry picked from commit aca0e26fe52eb8e36830686debf20b5fb9daf0b0)